Crate pallet_contract_operator

Modules

parameters

Structs

ContractHasOperator
A mapping from operated contract by operator to it.

ContractParameters
A mapping from contract to it's parameters.

Module

OperatorHasContracts
A mapping from operators to operated contracts by them.

Enums

Call
Dispatchable calls.

RawEvent
Events for this module.

Traits

ContractFinder

OperatorFinder

Trait
The module's configuration trait.

TransferOperator

Type Definitions

Event
RawEvent specialized for the configuration Trait